requirements.txt
Django==4.0
django-cors-headers==3.10.1
PyMySQL==1.0.2
bcrypt==3.2.0
gunicorn==20.1.0
ipython==7.30.0
mysqlclient==2.1.0
PyJWT==2.3.0
boto3==1.20.24
geopy==2.2.0
requests==2.26.0
python-dotenv=0.19.2